refactor(test): split the #3058 differential's CQLite-written fixture builders out (#3095, epic #1135)
The merge brought main's #3097 case (1355 -> 1482 lines) and our static work pushed it
to 1556, over the 1500-line test target, which the file-size ratchet FAILs. Split by
responsibility rather than waiving: the WriteEngine fixture builders and their
schema/DDL/timestamp constants move to tests/differential_fixtures/mod.rs, following the
crate's existing tests/fixture_support + tests/concurrent_support convention.

issue_3058_forced_path_differential.rs 1556 -> 1168 (below its pre-merge 1482 and below
the limit); differential_fixtures/mod.rs 409. The support module carries the
CQLite-written SCOPE reminder (its row content is never a Cassandra oracle — #3042) so
that caveat travels with the builders instead of living only at the call sites.

No behaviour change: the differential's full case list passes, including main's new
#3097 clustered_uncomp case and the collections/UDT fail-closed cases the fuller
/data/datasets corpus now exercises.
